refactor: extract ALE task material CLI commands
